requirements.txt
Pillow
numpy
scipy
matplotlib
opencv-python
pydot
imageio